storage files not zipped

+++ esofile_reader/storage/storage_files.py
@@ -155,37 +155,47 @@ class ParquetFile(BaseFile):
 
     @classmethod
     @profile(entries=10, sort="time")
-    def load_file(cls, source: Union[str, Path, io.BytesIO], dest_dir: Union[str, Path]):
+    def load_file(cls, source: Union[str, Path, io.BytesIO], dest_dir: Union[str, Path] = "") -> "ParquetFile":
         source = source if isinstance(source, (Path, io.BytesIO)) else Path(source)
+        if isinstance(source, io.BytesIO) or source.suffix == cls.EXT:
+            # extract content in temp folder
+            with ZipFile(source, "r") as zf:
+                # extract info to find out dir name
+                tempdir = tempfile.mkdtemp()
+                tempson = zf.extract("info.json", path=tempdir)
+                with open(Path(tempson), "r") as f:
+                    info = json.load(f)
+                shutil.rmtree(tempdir, ignore_errors=True)
 
-        # extract content in temp folder
-        with ZipFile(source, "r") as zf:
-            tempdir = tempfile.mkdtemp()
-            tempson = zf.extract("info.json", path=tempdir)
-            with open(Path(tempson), "r") as f:
+                # extract all the content
+                file_dir = Path(dest_dir, info["name"])
+                file_dir.mkdir()
+                zf.extractall(file_dir)
+
+        elif source.is_dir():
+            with open(Path(source, "info.json"), "r") as f:
                 info = json.load(f)
-            shutil.rmtree(tempdir, ignore_errors=True)
-            file_dir = Path(dest_dir, info["name"])
-            zf.extractall(file_dir)
+                file_dir = source
+        else:
+            raise IOError(f"Invalid file type loaded. Only '{cls.EXT}' files are allowed")
 
-            pqf = ParquetFile(
-                id_=info["id"],
-                file_path=info["file_path"],
-                file_name=info["file_name"],
-                file_created=datetime.fromtimestamp(info["file_created"]),
-                data=file_dir,
-                totals=info["totals"],
-                name=info["name"],
-                pardir=dest_dir,
-            )
+        pqf = ParquetFile(
+            id_=info["id"],
+            file_path=info["file_path"],
+            file_name=info["file_name"],
+            file_created=datetime.fromtimestamp(info["file_created"]),
+            data=file_dir,
+            totals=info["totals"],
+            name=info["name"],
+            pardir=file_dir.parent,
+        )
 
         return pqf
 
-    @profile(entries=10, sort="time")
-    def save_as(self, dir_: Union[str, Path] = None, name: str = None) -> Union[Path, io.BytesIO]:
-        """ Save parquet storage into given location. """
-        # store json summary file
-        self.data.save_info_parquets()
+    def save_meta(self) -> Path:
+        # store column, index and chunk table parquets
+        for tbl in self.data.tables.values():
+            tbl.save_info_parquets()
 
         # store attributes as json
         info = Path(self.workdir, f"info.json")
@@ -202,6 +212,13 @@ class ParquetFile(BaseFile):
                 "totals": self.totals,
             }, f, indent=4)
 
+        return info
+
+    @profile(entries=10, sort="time")
+    def save_as(self, dir_: Union[str, Path] = None, name: str = None) -> Union[Path, io.BytesIO]:
+        """ Save parquet storage into given location. """
+        info = self.save_meta()
+
         # use memory buffer if name or dir is not specified
         if name is None or dir_ is None:
             logging.info("Name or dir not specified. Saving zip into IO Buffer.")
@@ -214,7 +231,6 @@ class ParquetFile(BaseFile):
             zf.write(info, arcname=info.name)
             for dir_ in [d for d in self.workdir.iterdir() if d.is_dir()]:
                 for file in [f for f in dir_.iterdir() if f.suffix == ".parquet"]:
-                    arcname = file.relative_to(self.workdir)
-                    zf.write(file, arcname=arcname)
+                    zf.write(file, arcname=file.relative_to(self.workdir))
 
         return device
